refactor: remove mesos logic in uniapps.resource (#1414)
* bugfix: fix issue #1301

* refactor: remove mesos logic in uniapps.resource
