+++ bcs-app/backend/uniapps/resource/views/ingress.py
@@ -12,14 +12,12 @@ Unless required by applicable law or agreed to in writing, software distributed 
 an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
-import copy
 import logging
 
 from django.utils.translation import ugettext_lazy as _
 from rest_framework import viewsets
 
 from backend.apps import utils as app_utils
-from backend.apps.constants import ClusterType
 from backend.components.bcs import k8s
 from backend.templatesets.legacy_apps.configuration.k8s.serializers import K8sIngressSLZ
 from backend.templatesets.legacy_apps.instance.constants import K8S_INGRESS_SYS_CONFIG
@@ -27,12 +25,15 @@ from backend.uniapps.application.base_views import BaseAPI
 from backend.uniapps.application.utils import APIResponse
 from backend.uniapps.resource.views import ResourceOperate
 from backend.utils.errcodes import ErrorCode
-from backend.utils.error_codes import error_codes
 
 logger = logging.getLogger(__name__)
 
 
 class IngressResource(viewsets.ViewSet, BaseAPI, ResourceOperate):
+    category = 'ingress'
+    cate = 'K8sIngress'
+    sys_config = K8S_INGRESS_SYS_CONFIG
+
     def get_ingress_by_cluser_id(self, request, params, project_id, cluster_id):
         """查询configmaps"""
         access_token = request.user.token.access_token
@@ -47,22 +48,14 @@ class IngressResource(viewsets.ViewSet, BaseAPI, ResourceOperate):
 
     def get(self, request, project_id):
         """获取项目下的所有Ingress"""
-        project_kind = request.project.kind
-        project_kind_name = ClusterType.get(project_kind)
-
-        if project_kind_name != 'Kubernetes':
-            raise error_codes.CheckFailed(_("K8S项目才有Ingress"))
-
         cluster_dicts = self.get_project_cluster_info(request, project_id)
         cluster_data = cluster_dicts.get('results', {}) or {}
 
         # 获取命名空间的id
         namespace_dict = app_utils.get_ns_id_map(request.user.token.access_token, project_id)
 
-        s_cate = 'K8sIngress'
         is_decode = False
         params = {}
-        access_token = request.user.token.access_token
         data = []
         # 如果请求参数中cluster_id存在，根据cluster_id过滤集群信息
         query_cluster_id = request.query_params.get("cluster_id")
@@ -79,9 +72,7 @@ class IngressResource(viewsets.ViewSet, BaseAPI, ResourceOperate):
             self.handle_data(
                 request,
                 cluster_data,
-                project_kind,
-                s_cate,
-                access_token,
+                self.cate,
                 project_id,
                 cluster_id,
                 is_decode,
@@ -97,18 +88,11 @@ class IngressResource(viewsets.ViewSet, BaseAPI, ResourceOperate):
         return APIResponse({"code": ErrorCode.NoError, "data": {"data": data, "length": len(data)}, "message": "ok"})
 
     def delete_ingress(self, request, project_id, cluster_id, namespace, name):
-        self.category = 'ingress'
-        self.k8s_cate = 'K8sIngress'
         return self.delete_resource(request, project_id, cluster_id, namespace, name)
 
     def batch_delete_ingress(self, request, project_id):
-        self.category = 'ingress'
-        self.k8s_cate = 'K8sIngress'
         return self.batch_delete_resource(request, project_id)
 
     def update_ingress(self, request, project_id, cluster_id, namespace, name):
-        self.category = 'ingress'
-        self.k8s_cate = 'K8sIngress'
-        self.k8s_slz = K8sIngressSLZ
-        self.k8s_sys_config = K8S_INGRESS_SYS_CONFIG
+        self.slz = K8sIngressSLZ
         return self.update_resource(request, project_id, cluster_id, namespace, name)
